Create Room database and DAOs for Transaction, User, Category, and Budget entities
