--- app/src/main/java/com/example/expense_tracking_project/data/dataSource/empty.kt
@@ -1,4 +1,0 @@
-package com.example.expense_tracking_project.data.dataSource
-
-class empty {
-}
