+++ app/src/main/java/com/example/expense_tracking_project/presentation/vm/transaction_list/TransactionViewModel.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.viewModelScope
 import com.example.expense_tracking_project.data.dataSource.AppDatabase
-import com.example.expense_tracking_project.data.dataSource.Transaction
+import com.example.expense_tracking_project.data.dataSource.Transaction.Transaction
 import com.example.expense_tracking_project.domain.repository.Transaction.TransactionRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -55,15 +55,17 @@ class TransactionViewModel(application: Application) : AndroidViewModel(applicat
 
     fun hideTransaction(transaction: Transaction) {
         viewModelScope.launch(Dispatchers.IO) {
-            try {
-                val deletedTransaction = transaction.copy(
-                    isDeleted = true,
-                    updated_at = Date()
-                )
-                repository.update(deletedTransaction)
-                Log.d("TransactionDelete", "Soft delete applied to transaction.")
-            } catch (e: Exception) {
-                Log.e("TransactionDelete", "Error applying soft delete: ${e.message}", e)
+            if (!transaction.isDeleted) {
+                try {
+                    val deletedTransaction = transaction.copy(
+                        isDeleted = true,
+                        updated_at = Date()
+                    )
+                    repository.update(deletedTransaction)
+                    Log.d("TransactionDelete", "Soft delete applied to transaction.")
+                } catch (e: Exception) {
+                    Log.e("TransactionDelete", "Error applying soft delete: ${e.message}", e)
+                }
             }
         }
     }
